Image Based light for Diffuse ToneMapping now occours in colour correct. Code Clean Up

+++ GraphicsEngine/GraphicsEngine/Core/BaseWindow.h
@@ -53,7 +53,7 @@ protected:
 	class Editor_Camera* EditorCamera = nullptr;
 	class SceneJSerialiser* Saver = nullptr;
 	
-	float DeltaTime = 0.0f;
+	double DeltaTime = 0.0;
 	const float TickRate = 1.0f / 120.0f;
 
 	virtual void WindowUI() {};
